+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ##### 05.31.2024
 
+> Admin student registeration flow completion
 > Admin, lecturer login api integration and UI development
 > Admin 'add-admin' and 'add-lecturer' UI development and API integration
 > Course registration csv parsing, logic redefinition and API integration